Fix recursive casting of models
For example: passing `http_retry_properties` to `ClusterResource` did not end up in a `HTTPRetryProperties`, but in a dict - unlike the Pydantic model describes.

This is because we use `construct()`; a Pydantic method that does not validate input.

In turn, that has two reasons:

- Prevent performance impact of validation (we trust the data returned by the API; it is not necessary to be validated again)
- Work around Stoplight returning data not matching patterns (see https://github.com/stoplightio/prism/issues/2698)

Pydantic does not provide a way to skip validation while casting models recursively (see https://github.com/pydantic/pydantic/issues/3471 and https://github.com/pydantic/pydantic/issues/1168)

I attempted adding a custom factory method to `CoreApiModel`, which loops through fields on the model and casts input to them, while skipping validation (by calling `construct()`). However, the sheer amount of possible situations make that a hard to maintain solution: the input could be `__root__` (in case of a `RootModel`), or a list, or a dict.

Instead, I opted for a mockable solution: we now use `parse_obj`, which takes a dict rather than kwargs (so no functional difference), but calling `parse_obj` rather than the model's construct method (`BaseModel(**kwargs)`) - making it easy to mock in tests, effectively skipping validation and casting in tests, while doing so in real-world scenarios. This does mean that responses from the API are validated in the real world - which was one of the reasons to use `construct()` in the first place as described above - but that's a matter of elegancy, not a practical concern.